navigation with columns added

+++ query.py
@@ -15,10 +15,10 @@ supabase: Client = create_client(url, key)
 
 
 #@st.cache_data(ttl=3600)
-def load_portfolio():
+def load_portfolio(user_id):
  
     try:
-        response_all_stock_data = supabase.table("fet_portfolio_holdings").select("type","quantity","average_price","asset","symbol").eq("id", 56).execute()
+        response_all_stock_data = supabase.table("fet_portfolio_holdings").select("type","quantity","average_price","asset","symbol").eq("user_id", user_id).execute()
     except APIError as e:
         print(e)
         response = st.error("Error in Retrieving the data, Retry after sometime")
@@ -26,6 +26,17 @@ def load_portfolio():
 
     return pd.DataFrame(data_all_stock_data)
 
+
+def load_mf_transactions(user_id):
+ 
+    try:
+        response_all_mf_data = supabase.table("fet_portfolio_holdings_mf_transactions").select("id","fund_name","txn_type","amount","nav","units","created_at").eq("user_id", user_id).execute()
+    except APIError as e:
+        print(e)
+        response = st.error("Error in Retrieving the data, Retry after sometime")
+    mf_data= response_all_mf_data .data
+
+    return pd.DataFrame(mf_data)
 
 def insert_portfolio1(ins_data):
     return (
@@ -43,14 +54,28 @@ def get_mf_data(user_id,asset):
         .execute()
     )
 
-def update_portfolio(qty,avg_price,asset):
+def insert_mf_holdings(user_id,type,qty,avg_price,asset):
+    return (
+        supabase.table("fet_portfolio_holdings")
+        .insert({"user_id": user_id, "type": type,"asset": asset,"quantity": qty,"average_price":avg_price})
+        .execute()
+    )
+
+def insert_mf_transactions(mfs_data):
+    return (
+        supabase.table("fet_portfolio_holdings_mf_transactions")
+        .insert(mfs_data)
+        .execute()
+    )
+
+def update_portfolio(qty,avg_price,asset,user_id):
 
     try:
 
         response = (
             supabase.table("fet_portfolio_holdings")
             .update({"quantity": qty,"average_price":avg_price})
-            .eq("id", 56)
+            .eq("user_id",user_id)
             .eq("asset", asset)
             .execute()
         )
@@ -65,14 +90,14 @@ def update_portfolio(qty,avg_price,asset):
 
 
 
-def delete_portfolio(asset):
+def delete_portfolio(user_id,asset):
 
     try:
 
         response = (
             supabase.table("fet_portfolio_holdings")
             .delete()
-            .eq("id", 56)
+            .eq("user_id", user_id)
             .eq("asset", asset)
             .execute()
         )
@@ -84,3 +109,54 @@ def delete_portfolio(asset):
         response = st.error("Error in delete the data, Retry after sometime")
 
     return response
+
+
+def delete_mf_transaction(user_id,asset):
+
+    try:
+
+        response = (
+            supabase.table("fet_portfolio_holdings_mf_transactions")
+            .delete()
+            .eq("user_id", user_id)
+            .eq("fund_name", asset)
+            .execute()
+        )
+
+        print(response)
+
+    except APIError as e:
+        print(e)
+        response = st.error("Error in delete the data, Retry after sometime")
+
+    return response
+
+def delete_mf_transaction_id(user_id,asset,id):
+    try:
+
+        response = (
+            supabase.table("fet_portfolio_holdings_mf_transactions")
+            .delete()
+            .eq("user_id", user_id)
+            .eq("id", id)
+            .eq("fund_name", asset)
+            .execute()
+        )
+
+        print(response)
+
+    except APIError as e:
+        print(e)
+        response = st.error("Error in delete the data, Retry after sometime")
+
+    return response
+
+
+def get_mf_data_amount(user_id,asset):
+    return (
+        supabase.table("fet_portfolio_holdings")
+        .select("quantity","average_price")
+        .eq("user_id",user_id)
+        .eq("asset",asset)
+        .execute()
+    )
